requirements.txt
absl-py==1.4.0
anyio==3.7.0
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
deepface==0.0.79
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
flatbuffers==23.5.26
gast==0.4.0
gdown==4.7.1
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
imutils==0.5.4
itsdangerous==2.1.2
jax==0.4.12
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mtcnn==0.1.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
playsound==1.3.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
